+++ app/screens/addTransaction.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Text, View } from 'react-native';
+
+export const AddTransaction = () => {
+    return (
+        <View>
+            <Text>
+                Add here income and expenses
+            </Text>
+        </View>
+    )
+}