+++ app/screens/expenseGraph.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Text, View } from 'react-native';
+
+export const ExpenseGraph = () => {
+    return (
+        <View>
+            <Text>
+                Expenses will show in a graph for last 7 days
+            </Text>
+        </View>
+    )
+}