+++ app/screens/homescreen.js
@@ -1,9 +1,13 @@
 import React from 'react';
+// import { StyleSheet, Text, View } from 'react-native';
 import { Image, StatusBar, StyleSheet, Text, TouchableOpacity, View } from 'react-native';
 
 function Homescreen(props) {
     return (
         <View style={styles.container}>
+            {/* <Text>
+                Home
+            </Text> */}
             <StatusBar
                 backgroundColor="#FDFAF6"
                 barStyle="dark-content"
@@ -24,6 +28,12 @@ function Homescreen(props) {
                     <Text style={styles.spendingAmount}>- $320.60</Text>
                 </TouchableOpacity>
             </View>
+
+            <TouchableOpacity style={styles.addInputWrapper}>
+                <View style={styles.addMoney}></View>
+                <View style={styles.addExpense}></View>
+            </TouchableOpacity>
+
         </View>
     );
 }
@@ -33,7 +43,7 @@ const styles = StyleSheet.create({
         flex: 1,
         paddingTop: 10,
         paddingHorizontal: 10,
-        backgroundColor: "#FDFAF6",
+        // backgroundColor: "#595260",
     },
     userSectionWrapper: {
         paddingHorizontal: 10,
@@ -50,7 +60,7 @@ const styles = StyleSheet.create({
         letterSpacing: 2,
     },
     currentStatusWrapper: {
-        paddingTop: 10,
+        paddingTop: 40,
         flexDirection: "row",
         justifyContent: "space-around"
     },
@@ -86,6 +96,36 @@ const styles = StyleSheet.create({
     spendingAmount: {
         fontSize: 20,
         fontWeight: "bold"
+    },
+    addInputWrapper: {
+        paddingVertical: 15,
+        position: 'absolute',
+        bottom: 30,
+        right: 15,
+        borderRadius: 20,
+        // shadowColor: "#000",
+        // shadowOffset: {
+        //     width: 0,
+        //     height: 8,
+        // },
+        // shadowOpacity: 0.44,
+        // shadowRadius: 10.32,
+
+        // elevation: 16,
+        backgroundColor: "#FDF6F0"
+    },
+    addMoney: {
+        height: 50,
+        width: 50,
+        borderRadius: 20,
+        backgroundColor: "#055052",
+        marginBottom: 10,
+    },
+    addExpense: {
+        height: 50,
+        width: 50,
+        borderRadius: 20,
+        backgroundColor: "#B61919",
     },
 })