+++ app/screens/transaction.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import { Text, View } from 'react-native'
+
+export const Transaction = () => {
+    return (
+        <View>
+            <Text>
+                Transaction screen
+            </Text>
+        </View>
+    )
+}