+++ app/changelog.md
@@ -4,3 +4,7 @@
 
 - user image and name section added.
 - income and expenses status section added.
+
+> Chnagelog: v-0.1.2
+
+- money input section design added.